+++ settings.php
@@ -326,6 +326,22 @@ if ($ADMIN->fulltree) {
     $setting->set_updatedcallback('theme_reset_all_caches');
     $page->add($setting);
 
+    // Settings title for hiding the footer. We don't need a description here.
+    $name = 'theme_boost_campus/hidefooterheading';
+    $title = get_string('hidefooterheadingsetting', 'theme_boost_campus', null, true);
+    $setting = new admin_setting_heading($name, $title, null);
+    $page->add($setting);
+
+    // Hide the footer on the login page.
+    $name = 'theme_boost_campus/hidefooteronloginpage';
+    $title = get_string('hidefooteronloginpagesetting', 'theme_boost_campus', null, true);
+    $description = get_string('hidefooteronloginpagesetting_desc', 'theme_boost_campus', null, true);
+    $setting = new admin_setting_configcheckbox($name, $title, $description, 'no', 'yes', 'no' ); // Overriding default values
+    // yes = 1 and no = 0 because of the use of empty() in theme_boost_campus_get_pre_scss() (lib.php). Default 0 value would
+    // not write the variable to scss that could cause the scss to crash if used in that file. See MDL-58376.
+    $setting->set_updatedcallback('theme_reset_all_caches');
+    $page->add($setting);
+
     // Add tab to settings page.
     $settings->add($page);